correction of task 0

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *new_table;
-	unsigned int i;
+	unsigned long i;
 
 	if (size == 0 || size > ULONG_MAX)
 		return (NULL);
